+++ src/elements/conditions/NotificationCondition.php
@@ -1,0 +1,29 @@
+<?php
+/**
+ * Notifier plugin for Craft CMS
+ *
+ * Send custom Twig messages when Craft events get triggered.
+ *
+ * @author    Double Secret Agency
+ * @link      https://plugins.doublesecretagency.com/
+ * @copyright Copyright (c) 2021 Double Secret Agency
+ */
+
+namespace doublesecretagency\notifier\elements\conditions;
+
+use Craft;
+use craft\elements\conditions\ElementCondition;
+
+/**
+ * Notification condition
+ * @since 1.0.0
+ */
+class NotificationCondition extends ElementCondition
+{
+    protected function conditionRuleTypes(): array
+    {
+        return array_merge(parent::conditionRuleTypes(), [
+            // ...
+        ]);
+    }
+}